scripts/mirror-actors/pyproject.toml
[tool.poetry]
name = "mirror-actors"
version = "0.1.0"
description = ""
authors = ["samuelarogbonlo <samuel@chainsafe.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
slack-sdk = "^3.26.2"
boto3 = "^1.34.22"
pygithub = "^2.1.1"
pylint = "^3.0.3"
pylint-actions = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
load-plugins = "pylint_actions"
